refactor(grpc): 129 stage 2 - hello/superhero/greeter slices + platform/codegen under axis-1
Consumer-graph gate on ruling 6a produced an asymmetric split, now documented
as the rule: ISuperheroService is client-consumed (web-spa binds it via
CreateGrpcService against the grpc-contracts assembly) so it stays a
-contracts.cs; IHelloService has no consumer outside the grpc family so it
takes the -application.cs seam beside its impl. Codegen hosted service moves
to grpc/platform/codegen (first live platform tree outside web; planted-file
guard proof ran). Protos stay in grpc-server/protos - documented grammar
exemption. Gotcha recorded: global usings do not flow through ProjectReference
(Grpc.Core/System.ServiceModel added to grpc-application).

Implemented by the stage-1/2 subagent; gates re-run and committed by the
orchestrator after the agent stood down.
