+++ source/container-apps/grpc/features/hello/hello/hello-response-contracts.cs
