+++ source/container-apps/grpc/platform/codegen/protobuf-generation-hosted-service-server.cs
