+++ source/container-apps/grpc/projects/grpc-application/global-usings.cs
@@ -2,4 +2,8 @@
 // Project-wide using directives so individual files omit repeated imports.
 #endregion
 
+global using Grpc.Core;
+global using System.ServiceModel;
+
+// Solution usings
 global using TimeWarp.Mediator;
